+++ text.cpp
@@ -175,11 +175,11 @@ std::string format_commandline(int argc, char **argv) {
 
 // for jsonpull to call from C
 char *dtoa_milo(double val) {
-        std::string s = milo::dtoa_milo(val);
-        char *dup = strdup(s.c_str());
-        if (dup == NULL) {
-                perror("strdup");
-                exit(EXIT_FAILURE);
-        }
-        return dup;
+	std::string s = milo::dtoa_milo(val);
+	char *dup = strdup(s.c_str());
+	if (dup == NULL) {
+		perror("strdup");
+		exit(EXIT_FAILURE);
+	}
+	return dup;
 }